requirements.txt
dash
pymongo
sqlalchemy
python-dotenv
pyjwt
dash_bootstrap_components